Feat: credential Sign up, Sign in - JWT

+++ pages/api/post/delete.ts
@@ -1,6 +1,8 @@
 import { connectDB } from '@/utils/database';
 import { NextApiRequest, NextApiResponse } from 'next';
 import { ObjectId } from 'mongodb';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '../auth/[...nextauth]';
 
 export default async function handler(
   req: NextApiRequest,
@@ -10,11 +12,20 @@ export default async function handler(
     if (req.method === 'DELETE') {
       const client = await connectDB;
       const db = client.db('siyeol_blog');
-      await db
+      const session = await getServerSession(req, res, authOptions);
+      const post = await db
         .collection('blog_post')
-        .deleteOne({ _id: new ObjectId(req.body) });
-      res.status(200).json('삭제 성공');
-      return;
+        .findOne({ _id: new ObjectId(req.body) });
+      if (!post) {
+        throw new Error('post not found');
+      }
+      if (post.author === session?.user?.email) {
+        await db
+          .collection('blog_post')
+          .deleteOne({ _id: new ObjectId(req.body) });
+        res.status(200).json('삭제 성공');
+        return;
+      }
     }
   } catch (error) {
     res.status(500).json({ error: 'failed to delete' });
